Delete Question 3 at Knowledge Check Section
Delete Question 3 at Knowledge Check Section

There is not an answer to the deleted question.

The correct answer is to nest them but the cirriculum doesn't explaining how to do this. This page never mentions about using both <ul> and <ol> at the same time. 

Also, there is not an answer reference to the question. 

It should be deleted.